requirements.txt
requests==2.26.0
apache-airflow==2.9.1
python-dotenv